fix: vocalization of `alert` components with `Voice Over` announcement (#1491)
