+++ OUDS/Core/Components/Sources/Dialogs/Alert/Internal/AlertMessage/AlertMessageContent.swift
@@ -67,6 +67,9 @@ struct AlertMessageContent: View {
         }
         .padding(.vertical, theme.alert.spacePaddingBlock)
         .fixedSize(horizontal: false, vertical: true)
+        .onAppear {
+            VoiceOverUtils.announce(accessibilityLabel)
+        }
     }
 
     // MARK: - Helpers
